Move logic out of reducers
To stop potentialy useage of date object and consistency
